searchfilters: add methods to access the ...Filters and the real first content filter getFirstContentFilterItem()
Background: If the user (eg. NewPipe) wants to execute a search but does not set
any Filter we have a problem. Some services (eg. Bitchute) need to have
always a default set of sort/content filters otherweise they are not working.
As the SearchQueryHandler only pushes the user's selection (and as mentioned could be emtpty)
We can not use it to get the default ones. Therefore we change SearchQueryHandlerFactory
so we can have access to the default sort/content filters.